clone ui-plugin-find-users; adapt for inventory
            Episode II
      ATTACK OF THE CLONES

There is unrest in the Inventory
Module. Several thousand instance
records are hard to find from items
and holdings records they relate to.

This separation, fomented under the
leadership of the mysterious Count
DontHaveAPlugin, makes it difficult
to relate items and holdings records
to instances.

Now some dude from Nanananabooboo
has straight up cloned ui-plugin-
find-users, changed the remote,
created a new GitHub repo, and
stuffed inventory-related props
into the code that used to refer
to users. That's it. Pretty boring,
actually. It would be more fun
to watch reruns of the Galactic
Senate voting on CSPAN-2. Well,
maybe not...

Refs UIIN-217