UIOR-345 don't use shared components
It's a lovely idea to use shared components for the common plugin
architecture, but we want to make sure we share the right architecture.
We went down this road once already with `SearchAndSort`, which now
takes about 10,000 props, which is to say, we probably shared the wrong
architecture there. For now, we will just keep everything local.

Refs UIOR-345